Refactor: gameInfoForm 코드 분리
